backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-decouple==3.8
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
openpyxl==3.1.2
pandas==2.1.3
xlsxwriter==3.1.9
opencv-python-headless==4.10.0.84
pillow==10.4.0
numpy==1.26.4
onnxruntime==1.19.2
scikit-learn==1.3.2
reportlab==4.0.7
weasyprint==60.2
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1
redis==5.0.1
python-redis-lock==4.0.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
cryptography==41.0.7
celery==5.3.4
flower==2.0.1
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3
marshmallow==3.20.1
jsonschema==4.20.0
jinja2==3.1.2
